Did some more hardcode changes for the demo.
Put ownerLabels as Annie and Shrek. Fixed major to be able to display
majors which are 3 letters long appropriately. On MyGroups, only one
group can be selected between ApartOfGroups and OwnedGroups. Removed
some old comments.
BUG: If you select a group via the Map screen, go to a different screen,
and then go back to the map screen, it will take you to the group screen
instead of the Map screen.
When creating a group, be sure to use the group's initials, so it can be
found using search.